Refactor GTM
