+++ examples/starter/src/data/pages/meta$page-title.json
@@ -1,3 +1,3 @@
 {
-  "content": "Home 123"
+  "content": "BodilessJS Starter Site"
 }
