--- examples/starter/src/data/pages/meta$pagetitle.json
@@ -1,3 +1,0 @@
-{
-	"content": "BodilessJS Starter Site"
-}
