--- examples/starter/src/data/pages/meta$pagetype.json
@@ -1,3 +1,0 @@
-{
-	"content": "Home"
-}
